+++ sdk/storage/storage-queue/src/QueueClient.ts
@@ -250,6 +250,7 @@ export class QueueClient extends StorageClient {
       | string,
     options?: NewPipelineOptions
   ) {
+    options = options || {};
     let pipeline: Pipeline;
     if (credentialOrPipelineOrQueueName instanceof Pipeline) {
       pipeline = credentialOrPipelineOrQueueName;
@@ -274,10 +275,11 @@ export class QueueClient extends StorageClient {
         if (isNode) {
           const queueName = credentialOrPipelineOrQueueName;
           const sharedKeyCredential = new SharedKeyCredential(
-            extractedCreds.accountName,
+            extractedCreds.accountName!,
             extractedCreds.accountKey
           );
           urlOrConnectionString = extractedCreds.url + "/" + queueName;
+          options.proxy = extractedCreds.proxyUri;
           pipeline = newPipeline(sharedKeyCredential, options);
         } else {
           throw new Error("Account connection string is only supported in Node.js environment");